play asset reloading reproduction
A minimal project making it easy to reproduce the asset reloading issue.

The problem appears when implementing a typesafe provided workaround (for getting a non-play project depending on a play! framework project to see the assets of the play project), and then introducing the sbt-digest plugin to the build.

To reproduce the problem,
$> sbt
> project admin
> run

- now hit localhost:9000/assets/javascripts/badge.js
- add a line to badge js - console.log("Does this force a --RELOAD--?");
- Upon refresh, note that the application fully reloads:

"[info] Packaging /Users/patrick/github/play-asset-reloading/admin/target/admin-0.1-SNAPSHOT-assets.jar ...
[info] Done packaging.

--- (RELOAD) ---

[info] play - Application started (Dev)"

Then,
$> rm project/sbt-digest.sbt

And repeat the procedure.

Interestingly, there is only one pipeline stage in the project/Build.scala file - gzip.